+++ Cargo.toml
@@ -33,6 +33,7 @@ members = [
 	"polkadot/service",
 
 	"substrate/bft",
+	"substrate/cli",
 	"substrate/client",
 	"substrate/client/db",
 	"substrate/codec",
